+++ R/getCellContents.R
@@ -3,6 +3,7 @@
 #' \code{getCellContents} Gets information about container cell contents.
 #' @param coreApi coreApi object with valid jsessionid
 #' @param containerCellId container cell number as a string
+#' @param fullMetadata get full metadata, default is FALSE
 #' @param useVerbose  Use verbose communication for debugging
 #' @export
 #' @return RETURN returns a list $entity contains cell information, $response contains the entire http response
@@ -14,6 +15,7 @@
 #' logOut(login$coreApi)
 #' }
 #' @author Craig Parman info@ngsanalytics.com
+#' @author Natasha Mora natasha.mora@thermofisher.com
 #' @description \code{getCellContents} -  Gets information about a single container cell contents.
 
 
@@ -21,16 +23,23 @@
 
 
 getCellContents <-
-  function(coreApi, containerCellId, useVerbose = FALSE) {
+  function(coreApi,
+             containerCellId,
+             fullMetadata = FALSE,
+             useVerbose = FALSE) {
+
     # make sure containerCellNum is numeric
     containerCellId <- as.numeric(containerCellId)
 
     resource <- "CELL"
 
-    expansion <- switch(EXPR = substr(coreApi$semVer, 1, 1),
-      "2" = "?$expand=CONTENT($expand=IMPL_SAMPLE_LOT)",
-      "3" = "?$expand=CELL_CONTENTS($expand=SAMPLE_LOT)",
-      print("?$expand=CELL_CONTENTS($expand=SAMPLE_LOT")
+    case(
+      grepl("[0-2]+\\.[0-9]+\\.[0-9]+", coreApi$semVer) ~ {
+        expansion <- "?$expand=CONTENT($expand=IMPL_SAMPLE_LOT)"
+      },
+      grepl("[3-9]+\\.[0-9]+\\.[0-9]+", coreApi$semVer) ~ {
+        expansion <- "?$expand=CELL_CONTENTS($expand=SAMPLE_LOT)"
+      }
     )
 
     query <-
@@ -41,8 +50,11 @@ getCellContents <-
         expansion
       )
 
-    header <-
-      c("Content-Type" = "application/json;odata.metadata=full", Accept = "application/json")
+    if (fullMetadata) {
+      header <- c("Content-Type" = "application/json", "Accept" = "application/json;odata.metadata=full")
+    } else {
+      header <- c("Content-Type" = "application/json", "Accept" = "application/json")
+    }
 
     response <-
       apiGET(