+++ R/getEntityByName.R
@@ -5,7 +5,7 @@
 #' @param coreApi coreApi object with valid jsessionid
 #' @param entityType entity type to get
 #' @param name name of entity to get
-#' @param fullMetadata - get full metadata
+#' @param fullMetadata - get full metadata, default is FALSE
 #' @param useVerbose TRUE or FALSE to indicate if verbose options should be used in http
 #' @return returns a list $entity contains entity information, $response contains the entire http response
 #' @export
@@ -18,6 +18,7 @@
 #' }
 #' @author Craig Parman info@ngsanalytics.com
 #' @author Adam Wheeler adam.wheeler@thermofisher.com
+#' @author Natasha Mora natasha.mora@thermofisher.com
 #' @description \code{getEntityByName}  Get an entity by barcode from the Core LIMS using the ODATA API.
 
 
@@ -26,7 +27,7 @@ getEntityByName <-
   function(coreApi,
              entityType,
              name,
-             fullMetadata = TRUE,
+             fullMetadata = FALSE,
              useVerbose = FALSE) {
     query <- utils::URLencode(paste0("?$filter=Name eq '", name, "'"))