+++ tests/testthat/test-createSampleLot.R
@@ -8,8 +8,9 @@ test_that(paste("test createSampleLot() on:", env$auth), {
     data$sampleType,
     data$sampleBarcode,
     body = NULL,
+    fullMetadata = TRUE,
     useVerbose = verbose
   )
-
   expect_gt(httr::content(result$response)$CI_LOT_NUM, 0)
+  expect_true(!is.null(result$entity$`Id@odata.type`))
 })